moving to nodes_images
